moved benchmark into specific dir
